rename pages to componenets since webapp would be a one-pager
