+++ demo-web-app/frontend/src/components/UploadMonitor.jsx
